# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.analyzeMarkdownResponse:
    properties:
      counts:
        additionalProperties:
          type: integer
        type: object
      images:
        items:
          $ref: '#/definitions/utils.MarkdownImage'
        type: array
      total:
        type: integer
    type: object
...
  handler.batchConvertResponse:
    properties:
      results:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
  handler.batchMarkdownRequest:
    properties:
      files:
        items:
          $ref: '#/definitions/handler.markdownRequest'
        type: array
      target_config_id:
        type: integer
    type: object
...
  handler.convertMarkdownResponse:
    properties:
      changed:
        type: integer
      content:
        type: string
      filename:
        type: string
      record:
        $ref: '#/definitions/model.ConversionRecord'
      status:
        type: string
    type: object
  handler.createConvertTaskRequest:
    properties:
      files:
        items:
          $ref: '#/definitions/handler.markdownRequest'
        type: array
      target_config_id:
        type: integer
      task_type:
        type: string
    type: object
  handler.errorResponse:
    properties:
      error:
        example: 请求失败
        type: string
    type: object
...
  handler.markdownRequest:
    properties:
      content:
        type: string
      filename:
        type: string
      target_config_id:
        type: integer
    type: object
...
  handler.recordResponse:
    properties:
      record:
        $ref: '#/definitions/model.ConversionRecord'
    type: object
  handler.recordsResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/model.ConversionRecord'
        type: array
    type: object
...
  handler.taskCreateResponse:
    properties:
      results:
        items:
          additionalProperties: true
          type: object
        type: array
      task:
        $ref: '#/definitions/model.ConversionTask'
    type: object
  handler.taskDetailResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/model.ConversionRecord'
        type: array
      task:
        $ref: '#/definitions/model.ConversionTask'
    type: object
  handler.tasksResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/model.ConversionTask'
        type: array
    type: object
...
  model.ConversionRecord:
    properties:
      converted_content:
        type: string
      created_at:
        type: string
      details:
        items:
          $ref: '#/definitions/model.ConversionRecordDetail'
        type: array
      error_message:
        type: string
      id:
        type: integer
      image_count:
        type: integer
      original_filename:
        type: string
      source_picbed:
        type: string
      status:
        description: success/failed/processing
        type: string
      target_picbed:
        type: string
      task_id:
        type: integer
      user_id:
        type: integer
    type: object
  model.ConversionRecordDetail:
    properties:
      created_at:
        type: string
      error:
        type: string
      id:
        type: integer
      original_url:
        type: string
      record_id:
        type: integer
      status:
        type: string
      target_url:
        type: string
    type: object
  model.ConversionTask:
    properties:
      created_at:
        type: string
      ended_at:
        type: string
      error:
        type: string
      failed:
        type: integer
      id:
        type: integer
      message:
        type: string
      started_at:
        type: string
      status:
        type: string
      success:
        type: integer
      task_type:
        type: string
      total:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  utils.MarkdownImage:
    properties:
      alt:
        type: string
      picbed:
        type: string
      raw:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
  description: Markdown 图床批量转换平台 API
  title: PicBed Switcher API
  version: "1.0"
paths:
...
  /api/convert/analyze:
    post:
      consumes:
      - application/json
      parameters:
      - description: Markdown 内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.markdownRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.analyzeMarkdownResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 分析 Markdown 文档中的图片地址
      tags:
      - convert
  /api/convert/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: 批量转换请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.batchMarkdownRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.batchConvertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 批量执行 Markdown 文档转换
      tags:
      - convert
  /api/convert/local-batch:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 本地批量上传清单 JSON，结构为 localBatchManifest
        in: formData
        name: manifest
        required: true
        type: string
      - description: 本地图片文件，可按 manifest.images[].file_key 指定字段名上传；同一字段可提交多个文件
        in: formData
        name: images
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.batchConvertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 批量上传 Markdown 中引用的本地图片并替换地址
      tags:
      - convert
  /api/convert/process:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转换请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.markdownRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.convertMarkdownResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 执行单个 Markdown 文档转换
      tags:
      - convert
  /api/convert/records:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.recordsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取转换历史
      tags:
      - convert
  /api/convert/records/{id}:
    get:
      parameters:
      - description: 记录 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.recordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取转换历史详情
      tags:
      - convert
  /api/convert/tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tasksResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取转换任务列表
      tags:
      - convert
    post:
      consumes:
      - application/json
      parameters:
      - description: 转换任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createConvertTaskRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.taskCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建转换任务
      tags:
      - convert
  /api/convert/tasks/{id}:
    get:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.taskDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取转换任务详情
      tags:
      - convert
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"